avocadoserver/models.py: make Test a standalone entity
With more data pointing to a test, it makes sense not to simply refer
to a test as a given test_tag inside a job, but as a standalone entity.

Other entitities, such as test activity and test data (logs, images, etc)
will relate to individual test instances (which will then have a tag and
be a part of a job).

Signed-off-by: Cleber Rosa <crosa@redhat.com>