jobs API: allow to retrieve job either by numeric id or uniqueident
Numeric ID, aka PK, is the integer automatically assigned to the
underlying database record of a given model. For jobs, another way
to refer to them is to use their unique identifier, a uuid4 number
that is either explicitly set or automatically generated.

By using the API URI /jobs/<uniqueident/ it's possible to quickly
and effectively (resource wise) determine if a given job exists on
the server.

Signed-off-by: Cleber Rosa <crosa@redhat.com>