Component Isolation: port of Autotest's frontend database model
This is the beginning of Avocado's Component Isolation feature implementation,
whose foundation is the database model. Since no fault was found in the
Autotest version of the database model, this is a very direct port.

Besides adding new models, this adds an optional field to a test
that can (optionally) point to the recorded test environment the test
was run under.

Signed-off-by: Cleber Rosa <crosa@redhat.com>