CORS support: add middleware that adds CORS headers to every response
avocado-server is meant to be accessed by a number of clients, including
HTML+JS clients making AJAX requests. For obivous reasons, if the server
doesn't hint with the appropriate headers, the user agent (browser) will
refuse to process the response.

This is intentionally a very simple implementation, that gives the most
permissive access control setting in every single request. If we find
out that we need finer grained control, then we can implement that later.

Signed-off-by: Cleber Rosa <crosa@redhat.com>