Makefile: add file with build related targets
This Makefile adds a bunch of build related targets. It's similar to
the ones on the main Avocado repo, but with a couple of noteworthy
features:

1) It will build source tarballs, SRPM and RPM packages both from the
current HEAD, or from the latest tagged release. The requirements here
are that `$ python avocadoserver/version.py` matches a tag name in the
repo.

2) It will do all SRPM/RPM builds using mock, which is the right thing
to do and ensures the environment where the packages were built is
sound. This can be expanded in the future to allow the building of
other Tier-1 supported distros, such as EL6 and EL7.

Signed-off-by: Cleber Rosa <crosa@redhat.com>