unittests: follow Django conventions and recommendations
This includes dropping a bit of custom code in favor of Django's more
automatic handling of tests setup and tear down.

The most straightforward way of running the unittests is asking
Django's "manage" application to do it, so let's follow that too.

And finally the big class of model unittests has been broken down into
many classes, so that the test name themselves are easier to read and
grasp.

Signed-off-by: Cleber Rosa <crosa@redhat.com>