Job model: allow a job to be fetched by a short(er) ID
It's customary of git to be able to refer to any commit by an id that
can distinguish it from all others, using a shorter version of that
identifier.

We're trying to make things work similarly with job IDs, and a good
starting point is in the model layer itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>