Selftests: ported functional test to unittests
Django and rest framework actually allow us to test a live running
server and poking it with a very simple but capable client.

So, it's possible to have tests at the API level run like
unittests. This is a port, with some enhancements, of the old
functional tests to a unittest approach.

Changes from v1:
 * Add docstrings to API unittests

Signed-off-by: Cleber Rosa <crosa@redhat.com>